Feature/refactoring x2t converter

+++ DesktopEditor/xml/include/xmlutils.h
@@ -250,7 +250,7 @@ namespace XmlUtils
 	std::vector<XmlUtils::CXmlNode> oNodes;         \
 	if (node.GetNodes(name, oNodes))                \
 	{                                               \
-		int nCount = oNodes.size();                 \
+		size_t nCount = oNodes.size();              \
 		for (size_t i = 0; i < nCount; ++i)         \
 		{                                           \
 			XmlUtils::CXmlNode & oItem = oNodes[i]; \
